application layer finished

+++ rpg-legends/src/main/java/com/github/campaign_progression/domain/RoomFactory.java
@@ -3,7 +3,7 @@ package com.github.campaign_progression.domain;
 
 public class RoomFactory {
 
-    public Room createNextRoom(float battlechance) {
+    public Room createNextRoom(double battlechance) {
         //TODO: Add input to constructors?
         if (Math.random() < battlechance) {
             return new BattleRoom();
